+++ client/src/javascript/components/torrent-list/ActionBar.js
@@ -1,0 +1,120 @@
+import classnames from 'classnames';
+import {formatMessage, injectIntl} from 'react-intl';
+import React from 'react';
+import stringUtil from 'universally-shared-code/util/stringUtil';
+
+import Action from './Action';
+import Add from '../icons/Add';
+import EventTypes from '../../constants/EventTypes';
+import PauseIcon from '../icons/PauseIcon';
+import Remove from '../icons/Remove';
+import SettingsStore from '../../stores/SettingsStore';
+import SortDropdown from './SortDropdown';
+import StartIcon from '../icons/StartIcon';
+import StopIcon from '../icons/StopIcon';
+import TorrentActions from '../../actions/TorrentActions';
+import TorrentFilterStore from '../../stores/TorrentFilterStore';
+import TorrentStore from '../../stores/TorrentStore';
+import UIActions from '../../actions/UIActions';
+
+const METHODS_TO_BIND = [
+  'handleAddTorrents',
+  'handleRemoveTorrents',
+  'handleSortChange',
+  'handleStart',
+  'handleStop',
+  'handleSettingsChange'
+];
+
+class ActionBar extends React.Component {
+  constructor() {
+    super();
+
+    this.state = {
+      sortBy: SettingsStore.getFloodSettings('sortTorrents'),
+      torrentListViewSize: SettingsStore.getFloodSettings('torrentListViewSize')
+    };
+
+    METHODS_TO_BIND.forEach((method) => {
+      this[method] = this[method].bind(this);
+    });
+  }
+
+  componentDidMount() {
+    SettingsStore.listen(EventTypes.SETTINGS_CHANGE, this.handleSettingsChange);
+  }
+
+  componentWillUnmount() {
+    SettingsStore.unlisten(EventTypes.SETTINGS_CHANGE,
+      this.handleSettingsChange);
+  }
+
+  handleAddTorrents() {
+    UIActions.displayModal({id: 'add-torrents'});
+  }
+
+  handleRemoveTorrents() {
+    UIActions.displayModal({
+      id: 'remove-torrents'
+    });
+  }
+
+  handleSortChange(sortBy) {
+    this.setState({sortBy});
+    SettingsStore.saveFloodSettings({id: 'sortTorrents', data: sortBy});
+    UIActions.setTorrentsSort(sortBy);
+  }
+
+  handleStart() {
+    TorrentActions.startTorrents(TorrentStore.getSelectedTorrents());
+  }
+
+  handleStop() {
+    TorrentActions.stopTorrents(TorrentStore.getSelectedTorrents());
+  }
+
+  handleSettingsChange() {
+    this.setState({
+      sortBy: SettingsStore.getFloodSettings('sortTorrents'),
+      torrentListViewSize: SettingsStore.getFloodSettings('torrentListViewSize')
+    });
+  }
+
+  render() {
+    const classes = classnames(
+      'action-bar',
+      {
+        'action-bar--is-condensed':
+          this.state.torrentListViewSize === 'condensed'
+      }
+    );
+
+    return (
+      <nav className={classes}>
+        <div className="actions action-bar__item action-bar__item--sort-torrents">
+          <SortDropdown direction={this.state.sortBy.direction}
+            onSortChange={this.handleSortChange}
+            selectedProperty={this.state.sortBy.property} />
+        </div>
+        <div className="actions action-bar__item action-bar__item--torrent-operations">
+          <div className="action-bar__group">
+            <Action label="Start Torrent" slug="start-torrent" icon={<StartIcon />}
+              clickHandler={this.handleStart} />
+            <Action label="Stop Torrent" slug="stop-torrent" icon={<StopIcon />}
+              clickHandler={this.handleStop} />
+            <Action label="Pause Torrent" slug="pause-torrent" icon={<PauseIcon />}
+              clickHandler={this.handlePause} />
+          </div>
+          <div className="action-bar__group action-bar__group--has-divider">
+            <Action label="Add Torrent" slug="add-torrent" icon={<Add />}
+              clickHandler={this.handleAddTorrents} />
+            <Action label="Remove Torrent" slug="remove-torrent" icon={<Remove />}
+              clickHandler={this.handleRemoveTorrents} />
+          </div>
+        </div>
+      </nav>
+    );
+  }
+}
+
+export default injectIntl(ActionBar);