+++ config.template.js
@@ -4,6 +4,7 @@ const CONFIG = {
   dbPath: './server/db/',
   floodServerHost: '127.0.0.1',
   floodServerPort: 3000,
+  floodServerProxy: 'http://127.0.0.1',
   maxHistoryStates: 30,
   pollInterval: 1000 * 5,
   secret: 'flood',