+++ server/app.js
@@ -1,45 +1,45 @@
-'use strict';
-
 require('events').EventEmitter.defaultMaxListeners = Infinity;
 
 const torrentService = require('./services/torrentService');
 
-let bodyParser = require('body-parser');
-let compression = require('compression');
-let cookieParser = require('cookie-parser');
-let express = require('express');
-let favicon = require('serve-favicon');
-let morgan = require('morgan');
-let passport = require('passport');
-let path = require('path');
+const bodyParser = require('body-parser');
+const compression = require('compression');
+const cookieParser = require('cookie-parser');
+const express = require('express');
+const morgan = require('morgan');
+const passport = require('passport');
+const path = require('path');
 
-let app = express();
-let apiRoutes = require('./routes/api');
-let authRoutes = require('./routes/auth');
-let mainRoutes = require('./routes/main');
-let Users = require('./models/Users');
+const app = express();
+const apiRoutes = require('./routes/api');
+const authRoutes = require('./routes/auth');
+const paths = require('../client/config/paths');
+const Users = require('./models/Users');
+
+// Remove Express header
+if(process.env.NODE_ENV !== 'development') {
+  app.disable('x-powered-by');
+}
 
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'pug');
 
-// Remove Express header
-if(process.env.NODE_ENV !== 'development') app.disable('x-powered-by');
-
-// TODO: Add favicon...
-// app.use(favicon(__dirname + '/assets/favicon.ico'));
 app.use(morgan('dev'));
 app.use(passport.initialize());
 app.use(compression());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: false}));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'assets')));
 
 require('./config/passport')(passport);
 
 app.use('/api', apiRoutes);
 app.use('/auth', authRoutes);
-app.use('/', mainRoutes);
+
+// After routes, look for static assets.
+app.use(express.static(paths.appBuild));
+// After static assets, always return index.html
+app.use((req, res) => res.sendFile(path.join(paths.appBuild, 'index.html')));
 
 // Catch 404 and forward to error handler.
 app.use((req, res, next) => {
@@ -54,7 +54,8 @@ if (app.get('env') === 'development') {
     res.status(err.status || 500);
     res.render('error', {
       message: err.message,
-      error: err
+      error: err,
+      title: 'Flood Error'
     });
   });
 } else {
@@ -63,7 +64,8 @@ if (app.get('env') === 'development') {
     res.status(err.status || 500);
     res.render('error', {
       message: err.message,
-      error: {}
+      error: {},
+      title: 'Flood Error'
     });
   });
 }