+++ server/services/torrentService.js
@@ -1,9 +1,9 @@
 const deepEqual = require('deep-equal');
 const EventEmitter = require('events');
 
-const config = require('../../config.js');
 const clientRequestService = require('./clientRequestService.js');
 const clientRequestServiceEvents = require('../constants/clientRequestServiceEvents');
+const config = require('../../config.js');
 const formatUtil = require('../../shared/util/formatUtil');
 const methodCallUtil = require('../util/methodCallUtil');
 const notificationService = require('./notificationService.js');