+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   * Stream covers torrent list, transfer rate summary & history,
   torrent taxonomy, and notification count.
 * Close event stream after the window/tab has been inactive for 30 seconds
+* Refactor development experience, using `Webpack` & `WebpackDevServer`
+* Require users to build static assets again
 
 ## [1.0.0] (April 21, 2017)
 * First "official" release